+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:theme="@style/Theme.GlobalLogicApp"
         android:usesCleartextTraffic="true"
         tools:targetApi="m">
-        <activity android:name=".ui.MainActivity">
+        <activity android:name=".presentation.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
