+++ app/src/main/java/com/example/globallogicapp/di/component/ApplicationComponent.kt
@@ -1,8 +1,8 @@
 package com.example.globallogicapp.di.component
 
 import com.example.globallogicapp.di.module.ApplicationModule
-import com.example.globallogicapp.ui.DetailsFragment
-import com.example.globallogicapp.ui.ProductsFragment
+import com.example.globallogicapp.presentation.DetailsFragment
+import com.example.globallogicapp.presentation.ProductsFragment
 import dagger.Component
 import javax.inject.Singleton
