+++ app/src/main/res/layout/fragment_product.xml
@@ -8,7 +8,7 @@
     android:layout_marginLeft="16dp"
     android:layout_marginTop="10dp"
     android:layout_marginRight="16dp"
-    tools:context=".ui.ProductsFragment">
+    tools:context=".presentation.ProductsFragment">
 
     <androidx.cardview.widget.CardView
         android:id="@+id/empty_state"
