+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/productFragment"
-        android:name="com.example.globallogicapp.ui.ProductsFragment"
+        android:name="com.example.globallogicapp.presentation.ProductsFragment"
         android:label="fragment_product"
         tools:layout="@layout/fragment_product" >
         <action
@@ -16,7 +16,7 @@
     </fragment>
     <fragment
         android:id="@+id/blankFragment"
-        android:name="com.example.globallogicapp.ui.DetailsFragment"
+        android:name="com.example.globallogicapp.presentation.DetailsFragment"
         android:label="fragment_blank"
         tools:layout="@layout/fragment_details" >
         <argument
